Module mirrord_intproxy_protocol::codec

Custom codec used in layer <-> proxy communication. Supports both synchronous (required by the layer) and asynchronous (convenient for the proxy) IO. Asynchronous IO is feature-gated with the codec-async feature.

An encoded message consists of two parts:

prefix: 4 bytes containing payload length in bytes (big-endian u32)

payload: a value of a known type, encoded with bincode

Modules

codec_async 🔒

Structs

AsyncDecoder

Handles receiving messages of type T from the underlying AsyncRead of type W.

AsyncEncoder

Handles sending messages of type T through the underlying AsyncWrite of type W.

SyncDecoder

Handles receiving messages of type T from the underlying Write of type W.

SyncEncoder

Handles sending messages of type T through the underlying Write of type W.

Enums

CodecError

Errors that can occur when using this codec.

Constants

BUFFER_SIZE 🔒

Initial buffer size used by wrappers in this module. Buffers are used to encode and decode messages in memory.

PREFIX_BYTES 🔒

Length of the message prefix used by this codec. Determines the maximum length of the message.

Functions

make_async_framed

Creates a new pair of AsyncEncoder and AsyncDecoder, using the given asynchronous TcpStream.

make_sync_framed

Creates a new pair of SyncEncoder and SyncDecoder, using the given synchronous TcpStream.

Type Aliases

Result

Alias for Result type used by this codec.